refactor: dto의 역할 축소
